Add RX text and waterfall links
A double click on the waterfall inserts a text anchor than can be
clicked to return to that signal. Similarly, an Alt-click on the
waterfall scrolls the RX text to the most recent text anchor that
describes a similar modem and frequency combination.

If the "insert text on click" waterfall option is switched on,
text anchors will be inserted only if the click text box contains
the string "<FREQ>".

This implementation is based on an idea by Leigh, WA5ZNU.