Fix PSK viewer crash
The psk ctor was calling psk::restart() which in turn called the global
pskviewer's restart(), possibly deallocating memory in use by another psk
instance.  This would happen when changing PSK modes with the Viewer window
visible, as a new modem is constructed before the old one is destroyed.

Add a psk class viewpsk member and make the global pskviewer point to it.