Add PSK reporter activity text field
This patch adds a button and text field to the main window's rig control frame
that can be used to QSY to frequencies around which there have been recent
reception reports in a particular geographic area.  The area is determined by
the field text, which must be an IARU grid prefix.  If the field is empty, the
first two characters of the operator's locator are used.  If the locator is
unset, pskreporter.info uses (currently) IP geolocation to approximate it.

A popup menu is displayed when the user clicks the button or presses Return
inside the text field (which may be empty).  The popup shows frequencies sorted
by a measure of activity that gives more weight to transmissions.  If rig
control is enabled, a frequency can be sent to the rig by selecting it.

The data is retrieved from this URL:

  http://pskreporter.info/cgi-bin/psk-freq.pl

Or, with an non-empty text field:

  http://pskreporter.info/cgi-bin/psk-freq.pl?grid=TEXT