Lower slow CPU speed test threshold
The Intel Atom N270 and N280 processors score about 17 in the speed test
and are placed in the slowest category, which uses the LINEAR sample
rate converter.  However, they can easily handle the much better
SINC_FASTEST converter with some operator care.  Lower the slowest
threshold to 15 and assume that the operator will know not to run too
many CPU-intensive programs at the same time as fldigi.