Change waterfall focus handling
The waterfall widget steals focus when it has the mouse pointer and a
modifier key is pressed, and restores it when the modifier key is
released.  This allows is to handle certain keyboard shortcuts
(e.g. Shift/Ctrl+arrow keys) without requiring a click to take focus.

This commit changes the event handler to send non-arrow keys events to
the transmit text widget, so that things will mostly work even if the
user is typing into the transmit widget with the mouse pointer inside
the waterfall.